Changed the modifiers again.
I put Q on a weird spot as an experiment.  I did this in order so put mod3 on
its „usual“ spot.  I‘m not happy with it yet, but luckily I don‘t need the
letter very often, neither in english, nor in german.

On the plus side, I now have meta on a pleasant position on both sides.