The Q position wasn‘t working for me.
This also means that mod3 (caps lock) had to go somewhere else.

For now I put it back in the middle of the keyboard (lets see how that work for
programming).

Amend:

I updated the ASCII art in the code.  I also left the left CAPS-lock key intact,
I hope that this will not bite me.

I‘m still not happy with all the modifiers, especially control.  Also the space
key is surprisingly hard to use.